requirements.txt
openai==1.76.0
httpx==0.27.0
cryptography==43.0.1
PyJWT==2.9.0
fastapi==0.115.0
pydantic==2.4.2
uvicorn==0.22.0
python-dotenv~=1.1.0